techpost_rfc/pyproject.toml
[project]
name = "techpost-rfc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain",
    "langgraph",
    "langchain-openai",
    "langchain-community",
    "pydantic",
    "unstructured",
    "python-dotenv",
    "tiktoken"
]
